requirements.txt
# Snowmelt Control System - Python Dependencies

# MQTT client for Home Assistant integration
paho-mqtt>=1.6.1,<2.0

# Configuration file handling
PyYAML>=6.0

# GUI framework for touchscreen interface
PyQt5>=5.15.0

# Raspberry Pi GPIO - use rpi-lgpio for Pi 5 and Bookworm compatibility
# rpi-lgpio is a drop-in replacement for RPi.GPIO using libgpiod backend
rpi-lgpio>=0.6

# Optional: For development and testing
# pytest>=7.0
# pytest-cov>=4.0
